Use new tick function for total applications graph
Uses the calculateLinearTicks function to generate ticks for the total
applications graph. This allows all tick values to be passed into the
numberListFormatter that in turn allows it to better decide how to
format ticks. An example of a bug this fixes is for the following tick
values: [200, 400, 600, 800, 1000]. Based on just the extent the result
would be [0, 0, 1k, 1k, 1k] when it should be [0.2k, 0.4k, 0.6k, 0.8k,
1.0k].