Fix feature tests for sauce
Some feature tests included setup for a custom stub response. This is
not compatible with the way we run tests through Sauce as the app is
running in different process than the tests. Therefore, the response
setup was not being picked up. This can be solved by adding a route
to the stub controller that lets you add additional responses.
For now, we remove the instruction as we only ever use one response
for a given module.